requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
python-dotenv>=1.0.0
numpy>=2.0.0
pandas>=2.0.0
onnxruntime>=1.17.1
boto3>=1.28.85
streamlit>=1.28.1
pytest>=7.0.0
